+++ flecsi/util/array_ref.hh
@@ -51,54 +51,72 @@ struct span {
         std::data(std::declval<C &&>()))> (*)[],
       T (*)[]>>>
   constexpr span(C && c) : span(std::data(c), std::size(c)) {}
-
+  FLECSI_INLINE_TARGET
   constexpr iterator begin() const noexcept {
     return p;
   }
+
+  FLECSI_INLINE_TARGET
   constexpr iterator end() const noexcept {
     return q;
   }
 
+  FLECSI_INLINE_TARGET
   constexpr reverse_iterator rbegin() const noexcept {
     return reverse_iterator(end());
   }
+
+  FLECSI_INLINE_TARGET
   constexpr reverse_iterator rend() const noexcept {
     return reverse_iterator(begin());
   }
 
+  FLECSI_INLINE_TARGET
   constexpr reference front() const {
     return *begin();
   }
+
+  FLECSI_INLINE_TARGET
   constexpr reference back() const {
     return end()[-1];
   }
 
+  FLECSI_INLINE_TARGET
   constexpr reference operator[](size_type i) const {
     return begin()[i];
   }
 
+  FLECSI_INLINE_TARGET
   constexpr pointer data() const noexcept {
     return begin();
   }
 
   // FIXME: Spurious overflow for extremely large ranges
+  FLECSI_INLINE_TARGET
   constexpr size_type size() const noexcept {
     return end() - begin();
   }
+
+  FLECSI_INLINE_TARGET
   constexpr size_type size_bytes() const noexcept {
     return sizeof(element_type) * size();
   }
 
+  FLECSI_INLINE_TARGET
   constexpr bool empty() const noexcept {
     return begin() == end();
   }
 
+  FLECSI_INLINE_TARGET
   constexpr span first(size_type n) const {
     return {begin(), n};
   }
+
+  FLECSI_INLINE_TARGET
   constexpr span last(size_type n) const {
     return {end() - n, n};
   }
+  FLECSI_INLINE_TARGET
   constexpr span subspan(size_type i, size_type n = -1) const {
     return {begin() + i, n == size_type(-1) ? size() - i : n};
   }
@@ -305,31 +323,40 @@ struct iota_view {
 
   iota_view() = default;
   constexpr iota_view(I b, I e) : b(b), e(e) {}
-
+  FLECSI_INLINE_TARGET
   constexpr iterator begin() const noexcept {
     return b;
   }
+  FLECSI_INLINE_TARGET
   constexpr iterator end() const noexcept {
     return e;
   }
-
+  FLECSI_INLINE_TARGET
   constexpr bool empty() const {
     return b == e;
   }
+
+  FLECSI_INLINE_TARGET
   constexpr explicit operator bool() const {
     return !empty();
   }
 
+  FLECSI_INLINE_TARGET
   constexpr auto size() const {
     return e - b;
   }
 
+  FLECSI_INLINE_TARGET
   constexpr decltype(auto) front() const {
     return *begin();
   }
+
+  FLECSI_INLINE_TARGET
   constexpr decltype(auto) back() const {
     return *--end();
   }
+
+  FLECSI_INLINE_TARGET
   constexpr decltype(auto) operator[](I i) const {
     return begin()[i];
   }
@@ -616,6 +643,64 @@ FLECSI_TARGET transform_view(C &&, F)
 template<class C, class F>
 transform_view(const C &, F) -> transform_view<typename C::const_iterator, F>;
 /// \endcond
+
+/// A view of part of a range.  Analogous to a combination of
+/// \c std::take_view and \c std::drop_view from C++20.
+template<class R>
+struct substring_view {
+  using iterator = decltype(std::begin(std::declval<R>()));
+  using difference_type =
+    typename std::iterator_traits<iterator>::difference_type;
+
+  /// Select a substring of a range.
+  /// \param r input (stored in this object)
+  /// \param i starting index
+  /// \param n number of elements (must be in bounds)
+  constexpr substring_view(R r, difference_type i, difference_type n)
+    : alive(std::move(r)), b(std::next(std::begin(alive), i)), n(n) {}
+
+  FLECSI_INLINE_TARGET
+  constexpr iterator begin() const {
+    return b;
+  }
+  FLECSI_INLINE_TARGET
+  constexpr iterator end() const {
+    return std::next(b, n);
+  }
+
+  FLECSI_INLINE_TARGET
+  constexpr bool empty() const {
+    return !n;
+  }
+  FLECSI_INLINE_TARGET
+  constexpr explicit operator bool() const {
+    return n;
+  }
+
+  FLECSI_INLINE_TARGET
+  constexpr std::make_unsigned_t<difference_type> size() const {
+    return n;
+  }
+
+  FLECSI_INLINE_TARGET
+  constexpr decltype(auto) front() const {
+    return *b;
+  }
+  FLECSI_INLINE_TARGET
+  constexpr decltype(auto) back() const {
+    return b[n - 1];
+  }
+
+  FLECSI_INLINE_TARGET
+  constexpr decltype(auto) operator[](difference_type i) const {
+    return b[i];
+  }
+
+private:
+  R alive;
+  iterator b;
+  difference_type n;
+};
 
 /// \}
 } // namespace util