+++ doc/sphinx/src/news.rst
@@ -68,6 +68,10 @@ New features
   * Fields may be of non-portable types so long as they are used only by MPI tasks.
   * ``single`` accessors support ``->`` to access the members of a field value of class type.
 
+* Execution
+
+  * ``exec::trace`` improves performance of loops when used with the Legion backend.
+
 * Topologies
 
   * ``unstructured`` represents unstructured meshes.
@@ -81,8 +85,7 @@ New features
 
 * On-node parallelism
 
-  * ``parallel_for``, ``forall``, ``parallel_reduce``, and ``reduceall`` may be applied to subintervals of a range.
-  * These interfaces may be used without Kokkos enabled, in which case they run serially.
+  * ``parallel_for``, ``forall``, ``parallel_reduce``, and ``reduceall`` may be used without Kokkos enabled, in which case they run serially.
 
 * Utilities
 
@@ -92,6 +95,7 @@ New features
   * ``mdcolex`` accesses a multi-dimensional array using Fortran-like syntax.
   * ``mpi::one_to_alli`` scatters generated and/or serialized data with bounded memory usage.
   * MPI communication functions compute values to send in rank order and support mutable functors.
+  * ``substring_view`` represents part of another range.
 
 * Logging